// web/src/api/generated/model/signupIndex200One.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Les Canetons API
 * OpenAPI spec version: 0.0.1
 */
import type { SignupIndex200OneMenuTotals } from "./signupIndex200OneMenuTotals";
import type { SignupIndex200OneOccasion } from "./signupIndex200OneOccasion";
import type { SignupIndex200OneTablesItem } from "./signupIndex200OneTablesItem";

export type SignupIndex200One = {
  totalPersons: number;
  totalTables: number;
  menuTotals: SignupIndex200OneMenuTotals;
  tables: SignupIndex200OneTablesItem[];
  occasion: SignupIndex200OneOccasion;
};